+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * [9206](https://github.com/grafana/loki/pull/9206) **dannykopping**: Ruler: log rule evaluation detail.
 * [9184](https://github.com/grafana/loki/pull/9184) **periklis**: Bump dskit to introduce IPv6 support for memberlist
 * [9431](https://github.com/grafana/loki/pull/9431) **dannykopping**: Add more buckets to `loki_memcache_request_duration_seconds` metric; latencies can increase if using memcached with NVMe
+* [9466](https://github.com/grafana/loki/pull/9466) **appkins**: Azure: Create blob container if not found.
 
 ##### Fixes